# Preserve NAs when casting R data.frames to pandas.