@dfalbel 2023-08-10
This PR, will also cast `NA_character_` into `None` when converting R data.frames into Pandas data.frames. The old behavior would convert into the string `'NA'` possibly causing data loss.

With CRAN reticulate

``` r
library(reticulate)
df <- data.frame(
  string = c(NA, letters[1:4])
)
p <- r_to_py(df)
p
#>   string
#> 0     NA
#> 1      a
#> 2      b
#> 3      c
#> 4      d
p$string[0]
#> 'NA'
```

With this PR:

``` r
library(reticulate)
options(reticulate.pandas_force_numpy = TRUE)
df <- data.frame(
  string = c(NA, letters[1:4])
)
p <- r_to_py(df)
p
#>   string
#> 0   None
#> 1      a
#> 2      b
#> 3      c
#> 4      d
p$string[0]
#> None
```